Rename constants files.
